Changed rt types so that every pcap dlt has a matching rt type Added functions for conversion between rt type and pcap dlt format_rt now uses trace_set_err appropriately if rt_connect fails format_pcap will check if a trace is dead before trying to look at 	format data Dead traces now have their format_data set to NULL packet->type is now 32 bits, not 8

+++ lib/libtrace.h
@@ -111,7 +111,7 @@ typedef struct libtrace_packet_t {
 	void *buffer;		/**< allocated buffer */
 	size_t size;		/**< trace_get_framing_length()
 				 * +trace_get_capture_length() */
-	uint8_t type;		/**< rt protocol type for the packet */
+	uint32_t type;		/**< rt protocol type for the packet */
 } libtrace_packet_t;
 
 /** libtrace error information */
